third_party/blink/web_tests/platform/generic/fast/forms/reportValidity-valid-expected.txt

This test checks if reportValidity() returns correctly a true (meaning no error) result.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 lorem ipsum   
abc
Elements valid from the begining
PASS document.getElementsByTagName("fieldset")[0].reportValidity() is true
PASS document.getElementsByTagName("input")[0].reportValidity() is true
PASS document.getElementsByTagName("button")[0].reportValidity() is true
PASS document.getElementsByTagName("select")[0].reportValidity() is true
PASS document.getElementsByTagName("textarea")[0].reportValidity() is true
PASS document.getElementsByTagName("output")[0].reportValidity() is true
PASS document.getElementsByTagName("object")[0].reportValidity() is true
PASS document.getElementById("happy_form")[0].reportValidity() is true


Element valid when changed
PASS form.reportValidity() is true


PASS successfullyParsed is true

TEST COMPLETE
